rutas restricciones

+++ app/Http/Controllers/Admin/AppointmentController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\Appointment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+
 
 class AppointmentController extends Controller
 {
@@ -12,7 +14,8 @@ class AppointmentController extends Controller
      * Display a listing of the resource.
      */
     public function index()
-    {
+    {  
+        Gate::authorize('read_appointment');
         $appointments = Appointment::all();
         return view('admin.appointments.index', compact('appointments'));
     }
@@ -21,7 +24,7 @@ class AppointmentController extends Controller
      * Show the form for creating a new resource.
      */
     public function create()
-    {
+    {  Gate::authorize('create_appointment');
         return view('admin.appointments.create');
     }
 
@@ -29,7 +32,8 @@ class AppointmentController extends Controller
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
-    {
+    {    
+        Gate::authorize('create_appointment');
         $appointment = Appointment::create($request->all());
         return redirect()->route('appointments.index');
     }
@@ -38,7 +42,8 @@ class AppointmentController extends Controller
      * Display the specified resource.
      */
     public function show(Appointment $appointment)
-    {
+    { 
+        Gate::authorize('read_appointment');
         return view('admin.appointments.show', compact('appointment'));
     }
 
@@ -46,7 +51,8 @@ class AppointmentController extends Controller
      * Show the form for editing the specified resource.
      */
     public function edit(Appointment $appointment)
-    {
+    {  
+        Gate::authorize('update_appointment');
         return view('admin.appointments.edit', compact('appointment'));
     }
 
@@ -63,13 +69,15 @@ class AppointmentController extends Controller
      * Remove the specified resource from storage.
      */
     public function destroy(Appointment $appointment)
-    {
+    {  
+        Gate::authorize('delete_appointment');
         $appointment->delete();
         return redirect()->route('appointments.index');
     }
 
     public function consultation(Appointment $appointment)
-    {
+    {  
+        Gate::authorize('update_appointment');
         return view('admin.appointments.consultation', compact('appointment'));
     }
 }
