+++ app/Http/Controllers/Admin/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\User;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
 
@@ -13,7 +14,8 @@ class UserController extends Controller
      * Display a listing of the resource.
      */
     public function index()
-    {
+    {  
+        Gate::authorize('read_user');
         return view('admin.users.index');
     }
 
@@ -21,7 +23,9 @@ class UserController extends Controller
      * Show the form for creating a new resource.
      */
     public function create()
-    {
+
+    {  
+          Gate::authorize('create_user');
         $roles = Role::all();
 
         return view('admin.users.create', compact('roles'));
@@ -31,7 +35,8 @@ class UserController extends Controller
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
-    {
+    {   
+          Gate::authorize('create_user');
         $data = $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|string|email|max:255|unique:users',
@@ -70,7 +75,8 @@ class UserController extends Controller
      * Display the specified resource.
      */
     public function show(User $user)
-    {
+    {  
+          Gate::authorize('read_user');
         return view('admin.users.show', compact('user'));
     }
 
@@ -78,7 +84,8 @@ class UserController extends Controller
      * Show the form for editing the specified resource.
      */
     public function edit(User $user)
-    {
+    {   
+          Gate::authorize('update_user');
         $roles = Role::all();
 
         return view('admin.users.edit', compact('user', 'roles'));
@@ -88,7 +95,8 @@ class UserController extends Controller
      * Update the specified resource in storage.
      */
     public function update(Request $request, User $user)
-    {
+    {   
+         Gate::authorize('update_user');
         $data = $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|string|email|max:255|unique:users,email,'.$user->id,
@@ -126,7 +134,8 @@ class UserController extends Controller
      * Remove the specified resource from storage.
      */
     public function destroy(User $user)
-    {
+    {  
+         Gate::authorize('delete_user');
         $user->delete();
         session()->flash('swal',
             [
